Dev: Start work on instance upgrading (incomplete)
