+++ crates/ql_mod_manager/src/store/mod.rs
@@ -1,10 +1,11 @@
 use std::{
     collections::HashSet,
+    path::PathBuf,
     sync::{Arc, mpsc::Sender},
 };
 
 use chrono::DateTime;
-use ql_core::{GenericProgress, Instance, Loader, do_jobs, pt};
+use ql_core::{GenericProgress, Instance, Loader, do_jobs, err, pt};
 
 mod add_file;
 mod curseforge;
@@ -283,4 +284,63 @@ pub async fn get_download_link(
         ModId::Modrinth(n) => ModrinthBackend::get_download_link(instance, n, query_type).await,
         ModId::Curseforge(n) => CurseforgeBackend::get_download_link(instance, n, query_type).await,
     }
+}
+
+pub async fn get_locally_installed_mods(
+    dot_mc: PathBuf,
+    blacklist: HashSet<String>,
+    project_type: QueryType,
+) -> HashSet<LocalMod> {
+    let dirs: &[&str] = match project_type {
+        QueryType::Mods => &["mods"],
+        QueryType::ResourcePacks => &["resourcepacks", "texturepacks"],
+        QueryType::Shaders => &["shaderpacks"],
+        QueryType::DataPacks => &["datapacks"],
+        QueryType::ModPacks => return HashSet::new(),
+    };
+    let mut set = HashSet::new();
+
+    for dir in dirs {
+        let mods_dir_path = dot_mc.join(dir);
+
+        let mut dir = match tokio::fs::read_dir(&mods_dir_path).await {
+            Ok(dir) => dir,
+            Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
+                continue;
+            }
+            Err(err) => {
+                err!("While reading {dir} directory: {err}");
+                continue;
+            }
+        };
+
+        while let Ok(Some(entry)) = dir.next_entry().await {
+            let path = entry.path();
+            let Some(file_name) = path.file_name().and_then(std::ffi::OsStr::to_str) else {
+                continue;
+            };
+            if blacklist.contains(file_name) {
+                continue;
+            }
+            if let Ok(f) = entry.file_type().await {
+                if f.is_dir() {
+                    if project_type == QueryType::Mods {
+                        continue;
+                    }
+                } else {
+                    let Some(extension) = path.extension() else {
+                        continue;
+                    };
+                    if !(extension.eq_ignore_ascii_case("jar")
+                        || extension.eq_ignore_ascii_case("zip")
+                        || extension.eq_ignore_ascii_case("disabled"))
+                    {
+                        continue;
+                    }
+                }
+            }
+            set.insert(LocalMod(Arc::from(file_name), project_type));
+        }
+    }
+    set
 }
